requirements.txt
google-cloud-discoveryengine>=0.11.0
google-auth>=2.17.0
google-api-python-client>=2.100.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
